* commit 'd4f7f288a2a7e15f813cdaed1efc55953f837679':
  убрал штуки типа "юзер как юзер"
  небольшая правка
  поправил тесты
  сделал индексеры целочисленнымии

+++ replay/models/word2vec.py
@@ -84,7 +84,7 @@ class Word2VecRec(Recommender):
         self.vectors = (
             word_2_vec.fit(log_by_users)
             .getVectors()
-            .select(sf.col("word").cast("float").alias("item"), "vector")
+            .select(sf.col("word").cast("int").alias("item"), "vector")
             .cache()
         )
